change the type to uint8 for some fields

+++ cmd/broker/main.go
@@ -58,7 +58,7 @@ const (
 )
 
 const (
-	_ uint32 = iota
+	_ uint8 = iota
 	TypeHTTP
 	TypeTCP
 )
